:recycle: Refactor: Custom Hook 상태 로직 추가/상태 로직 없는 함수 utils로 이동(#120)
