--- src/hooks/useUploadImg.ts
@@ -1,11 +1,0 @@
-import { getDownloadURL, ref, uploadBytesResumable } from 'firebase/storage';
-
-import { storage } from '../firebase/config.ts';
-
-export default async function useUploadImg(path: string, file: File) {
-  const storageRef = ref(storage, path);
-  const uploadTask = uploadBytesResumable(storageRef, file);
-
-  const downloadURL = await getDownloadURL(uploadTask.snapshot.ref);
-  return downloadURL;
-}
